+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for CPU profiling
 - Add V8 build for Apple Silicon
 - Add support for throwing an exception directly via the isolate's ThrowException function.
+- Support for compiling a context-dependent UnboundScript which can be run in any context of the isolate it was compiled in.
+- Support for creating a code cache from an UnboundScript which can be used to create an UnboundScript in other isolates
+to run a pre-compiled script in new contexts.
 
 ### Changed
 - Removed error return value from NewIsolate which never fails